add: python2 support
* replace __str__ with .get_text()
* use tox for unit testing.

+++ src/inscriptis/css.py
@@ -100,7 +100,7 @@ class CssParse(object):
         unit = m.group(2)
 
         if unit not in ('em', 'qem', 'rem'):
-            value = round(value/8)
+            value = int(round(value/8))
         return value
 
     # ------------------------------------------------------------------------
